+++ src/components/Top.tsx
@@ -7,7 +7,7 @@ import { Helmet } from "react-helmet";
 import { AppState } from '../store'
 import TopSection from './common/TopSection'
 import WorkCard from './ui-parts/WorkCard'
-import { IWorkState } from '../states/states';
+import { IWorkState } from '../states/firebase';
 import { userActions } from '../actions/actions';
 import OctetttrussSvg from './asset/OctetttrussSvg'
 
@@ -101,7 +101,7 @@ const Top: React.FC<TopProps>  = (props: TopProps) => {
 
 function mapDispatchToProps(dispatch: Dispatch<Action<IWorkState | null>>) {
     return {
-        selecteWork: (work: IWorkState | null) => dispatch(userActions.SelecteWork(work)),
+        selecteWork: (work: IWorkState | null) => dispatch(userActions.selecteWork(work)),
     }
 }
